Removed the string[] as a type for entityid in conditions
First of all, love your vs extension!

I tried to add a condition to an automation with multiple entityid's (as a list), it parsed in your extension in vs code, but when running the HASS validation tool it failed. As I read in this issue:

https://github.com/home-assistant/core/issues/3620

It seems to be impossible to add multiple entityid's in a condition.